Worked on scoring
Improved the movement, added compatibility with other planets for different gravity ratings. Added a high score file so high scores can we tracked over time. Massive bugs with scoring feature at the moment, next time I work on it I will improve the scoring feature to work and then display the score on the screen as well as the current high score in the top corner. Then its onto losing and finally adding enemies to avoid.